Relicence Playslave (C++ version).
This is in preparation for potential usage of parts of playslave in
baps-playout, which is MIT licenced.

I believe that I/URY are the only copyright holder of this code, and
thus that this relicence is sound.

Note that the change to MIT complicates distribution of Playslave
with a GPL'd ffmpeg or libav.