Use chrono for timestamps.
This is very verbose and requires some hoop jumping, but is
somewhat more clear in its intent (and allows templates for
easily implementing 'seek X seconds').  Is it worth it?
Probably not.  Who knows.

Also, the seek timer is now marked as dirty whenever a new song
is loaded, a new seek is done, and on startup.