Add some Visual Studio field notes.
LibSoX makes the VS build a bit more complicated, so I've added these into the
tree to compensate.

These are slightly cobbled together from memory, and may be a bit incorrect or
incomplete in areas.  If people could try out VS builds and fix up issues in
the notes, this would be great.