Add TimeParser test.
This is not a complete test suite, as it does not check exceptions.  Currently,
I'm unsure how to use Catch for exception tests, so we only check the 'valid'
cases.

The RHS data has `ull` suffixes due to overflows when previous iterations of
the unit tests were run on Travis.