Rationalise Audio interfaces.
This is a big refactoring that cleans up the interface for Audio and
PlayerFile, makes Audio into an abstract class (implemented by the new
PipeAudio class), and generally makes things less confusing.

This has NOT been tested yet.  I expect a lot of audio code is broken,
and this should NOT be pushed in before it's been properly tested.  I
intend to write some unit tests to test the more basic properties, but
an actual demo run is needed.