Use libsndfile for oggs and wavs.
This is a bit of a cop-out, as we were going to use libvorbisfile for oggs.
However, libsndfile adds support for several other formats, like wav etc.

Why not libvorbisfile?  Because libvorbisfile makes us deal with things like
changing sample rates mid-track (!!).  Rather than dramatically refactor playd
to cope with this, we'll just use libsndfile, eh?

A possible thing to fix up in this commit is an implicit assumption that int is
32 bits.