sUbstitute SDL2 for PortAudio.
SDL2 is a more commonplace library, seems to fix some weird sound
behaviour on FreeBSD, is slightly cleaner in terms of APIs, and has
a 2 in it so it must be good.  It's also a somewhat round-about way
of getting rid of the PortAudio C++ bindings, which have been a pain
in the posterior to build for aeons.

Some changes to the Makefile/config have been made, most to accommodate
this, and some to make my life easier :3