Retire PlayerState.
All state information is now taken from Audio.  Currently this is
in a less than smooth manner, but this may be cleaned up afterwards.

A side-effect is that Play(), Stop(), and Eject() are now idempotent,
and succeed without doing anything if called unnecessarily.  I'd
like to keep this 'feature', as it simplifies code on both playd
and its clients' front.