Implement ID-based I/O addressing.
Now, whenever something is to be sent to the I/O system, it takes the
IoCore as its ResponseSink, and an additional integer uniquely
representing the connection in the IoCore.  These IDs are retained by
connections until they disconnect, in which case they are put up for
recycling.

Hopefully, this should stop proliferation of pointers to Connections,
which are not safe to leak outside IoCore because they can be destructed
at any point.