Send a quit acknowledgement before closing.
This has required a bit of a refactor to get it to work.  We no longer
explicitly close the libuv loop, but instead make sure that the server
and timer are closed when the Player quits, and each connection is
closed when, and only when, it receives an acknowledgement.

The current ack is 'STATE Quitting', which I think is the spec-standard
one, but this might want to be changed in the spec later as the
semantics differ somewhat from other uses of STATE.