Get building on VS2015.
This mainly involved removing the SDL2 nuget, getting SDL2 from /lib and /include, and otherwise mauling the vcxproj.

We can't get SDL2 from nuget because, in order to compile debug versions, we need to build SDL2 for debug.  This means rolling our own.

At some point we should probably get this so it reads debug and release libs from separate places---but release is broken at the moment anyway, I think.