Initial work on protocol3 playd.
This converts everything except the man page and tests to use and
discuss the draft third Bifrost protocol.

As can be seen, implementation of this is fairly simple, and I think
having a concrete implementation will aid discussion of it as a protocol
going forward.

There are some differences between this and the last draft of protocol3:

- quit is still available in playd, and will probably exist until we
implement control-C trapping;
- ACKs are not in the correct format, and I may need to go back and
revise the ACK format in the draft.