Merge CommandResult into Response.
CommandResults have no use other than to spawn Responses, and this can
be done using ::Success, ::Failure and ::Invalid anyway.  Those three
are moved into Response; IO is changed to deal with Responses coming out
of commands directly; and s/CommandResult/Response/ everywhere in the
code base.

I will fix the unit tests when I'm out of the hell that is Visual
Studio.