Refactoring to align to C++17 and C++ Core Guidelines (#126)
This is a fairly hefty stream of consciousness bundle that bundles a large amount of refactoring, including but not limited to:

-   Adopting C++17 and using some C++17 language and stdlib features;
-   Starting to realign naming conventions against the C++ Core Guidelines;
-   Making the travis build use a more modern and mostly off-the shelf build environment (this has resulted in me stripping out Coveralls integration, but I don't think this'll be a blocker);
-   Moved everything into namespaces (playd mostly, but playd::audio for audio code and playd::tests for tests);
-   Replaced the PortAudio ring buffer with a hand-written but fairly lightweight C++ implementation;
-   Various code-cleanup changes (method extractions, constant extractions, etc).

Eventually I want to move to std::filesystem for file paths, but I'll wait until macOS 10.15 comes out so that it'll work with AppleClang.

I apologise for the sheer amount of changes in this one PRQ: it's been a side-project I've repeatedly dropped, picked back up, dropped again, forgotten about, left to rot on various computers, and recently picked back up again.

* Copy in current master of Microsoft/GSL.

* Replace PortAudio ringbuffer with native C++ one

This ringbuffer is in native C++, and is probably slower than the
original but a damn sight nicer if I may say so myself.  May have odd
concurrency bugs.

* Fix undefined behaviour in tokeniser.

Oops... this was trying to get isspace of non-ASCII characters in a
nonportable way.

* Use spans for ringbuf operations.

Yay!

* A few minor changes.

* update gsl to 9d13cb

* I.4: use chrono for microseconds

At the same time, separate CanAnnounce... from actually updating
the last time, making most of the Announce logic pure as per F.8.

* SF.1: use h for header files

This will probably be controversial, but I thought since we're
doing a Core Guidelines blitz we might as well head for as much
compliance as possible.

catch.hpp is left as is for now.

* More renaming and core guidelines changes

Generally, moving from SnakeCase to Capitalised_camel_case on types, and
fixing some broken documentation.

* Fix a few bugs and typoes

Also change the Makefile a bit to keep it up to date.

* Update GSL and relocate

The GSL now expects to be absolutely referenceable on the compiler
include path, so some changes to the build systems needed to be
made.

* Fix conversion errors in test

This should unstick the classic build on Travis.

* Clang-format everything

Make things neat and clean!

* Get playd compiling on VS 2017

- Fix incorrect version check in cmake
- Paper over weirdness with std::make_unique by indirection
  (any better ways to do this?)
- Build and test, everything seems to be according to keikaku*

* TL Note: keikaku means plan

* Finally commit archaic GSL updates

I've forgotten what the rationale for these was.  Sorry...!

* Replace Catch with Catch2

This fixes some breakages when we move to C++17.

* Use C++17

We now require cmake3.11, to use the target_compile_features feature.

This hasn't been tested thoroughly on pre-C++17 compilers, so any such testing would be swell.

* Start adding some C++17 features

Nothing particularly exciting yet.

* Get rid of std::tie

C++17!

* Fix broken clang-format line

* Update GSL

* Base errors on top of std::exception

* Extract method StateResponseCode

* Clang-format and revert config changes

* Extract method DumpFileInfo

* clang-format9

* Try to get Travis to build (without coveralls)

If anyone wants to make coveralls work again, be my guest =3

* Fix missing vector header

* Begin the endless task of travis tweaking

* Delete POSIX makefile

`cmake` is now the only build system.  *sheds single, solitary tear*

I haven't deleted the ports, but I also haven't checked to see whether
they actually work with cmake.  They'll need some checking to see if
they still work.

* Make minor cleanup changes

These are all fairly minor refactors.  I'll do some more invasive stuff,
like adding string_views, separately.

* Fix inadvertent exception mess

- Makes exceptions a subclass of Runtime_error, so that `what()` gets
  the same message as `Message` and catch can return messages (TODO:
  clean this up a little)
- Drops a cast to `Error` in `Null_audio` that broke the catch in
  `Player`.

* Player: more cleanup

* main: string-viewify and fix related bugs

Also, as a bonus, make playd actually say why a device ID is invalid.

* Try nudge Travis into using higher stdlib

* mp3: factor out encoding mapping

* Use std::byte for things that expect bytes

This is mostly the ringbuffer, but also things that touch the
ringbuffer.

There's a bit of clang-format bleeding in this commit, for which I
apologise.

* Add namespaces and frantically try to reformat

Everything is now in the playd::, playd::audio, or playd::tests:: namespace.

Further namespace adjustment is forecoming.

* Bump column limit to 120

There aren't that many long lines in playd, and most of the breaks
that are being generated by clang-format are arbitrary/ugly, so I
_think_ on balance this is an improvement?

* Io: move to playd::io namespace

This means playd::Io_core is now just playd::io::Core.

* Clean up tests

Propagate some renames that got a bit lost along the way, collapse
weirdly line-broken artefacts, etc.

* Replace GSL with submodule

This might need a few prods to get Travis to work.  Not sure.

* Rename things back to CamelCase

Per @LordAro's suggestion.

This commit tries to change some things back to CAPS_SNAKE_CASE, but I
need to do a proper pass of checking this, as there are some
const/constexpr things where it's not entirely sure which case it should
be.