Add POS broadcasts on change of playing state
When `play`ing or `stop`ping, playd will now broadcast the exact
current position in the file after effecting the state change.
This makes for somewhat more precise resolution of where the
file was when the change occurred.

Fixes #128.

Needs unix testing, which I'll do on my Mac in a sec.